+++ source/parsing_identifiers/print_string.c
@@ -6,7 +6,7 @@
 /*   By: phemsi-a <phemsi-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/08 14:49:08 by phemsi-a          #+#    #+#             */
-/*   Updated: 2021/03/09 10:39:40 by phemsi-a         ###   ########.fr       */
+/*   Updated: 2021/03/09 18:33:18 by phemsi-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	print_string(flags *flag, va_list args)
 {
 	char	*string;
-	size_t	length;
+	int		length;
 
 	if (flag->precision)
 		string = ft_substr(va_arg(args, char *), 0, flag->precision);
